requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
sqlalchemy>=2.0.0
openai>=1.30.0
crewai>=0.28.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
fpdf2>=2.7.0
reportlab>=4.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
xgboost>=2.0.0
scikit-learn>=1.4.0
pypdf>=4.0.0
python-docx>=1.1.0